rename is_parallelzable to is_naive_parallelizable
