+++ src/transformers/models/gpt2/modeling_gpt2.py
@@ -373,7 +373,7 @@ class GPT2PreTrainedModel(PreTrainedModel):
     config_class = GPT2Config
     load_tf_weights = load_tf_weights_in_gpt2
     base_model_prefix = "transformer"
-    is_parallelizable = True
+    is_naive_parallelizable = True
     supports_gradient_checkpointing = True
 
     def __init__(self, *inputs, **kwargs):
