+++ src/transformers/models/t5/modeling_t5.py
@@ -714,7 +714,7 @@ class T5PreTrainedModel(PreTrainedModel):
     config_class = T5Config
     load_tf_weights = load_tf_weights_in_t5
     base_model_prefix = "transformer"
-    is_parallelizable = True
+    is_naive_parallelizable = True
     supports_gradient_checkpointing = True
 
     @property
